# faithgpt-platform/api/openapi.yaml
openapi: 3.0.3
info:
  title: FaithGPT API
  version: "1.0.0"
  description: >
    Canonical REST API contract for FaithGPT™ ("Transforming Scripture into Daily Living").
    Conventions, envelopes, pagination, streaming, and rate-limiting are documented narratively
    in /faithgpt-platform/docs/09-api-architecture.md — this document is the reviewable
    machine-readable snapshot. Schema names mirror /faithgpt-platform/schema/schema.prisma
    verbatim (see /faithgpt-platform/docs/00-shared-reference.md §13).
  contact:
    name: FaithGPT Engineering
    url: https://faithgpt.app
servers:
  - url: https://api.faithgpt.app/api/v1
    description: Production
  - url: https://staging-api.faithgpt.app/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# =====================================================================================
# PATHS
# =====================================================================================
paths:
...
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization detail
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOrganization'
        "404":
          $ref: '#/components/responses/NotFound'
        "403":
          $ref: '#/components/responses/Forbidden'
    patch:
      tags: [Organizations]
      summary: Update organization (name, seatLimit, billingEmail) — ADMIN role in org
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                seatLimit: { type: integer }
                billingEmail: { type: string, format: email, nullable: true }
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeOrganization'
...
  /devotions/generate:
    post:
      tags: [Devotions]
      summary: Generate a devotion via the AI Verse-to-Devotion Engine™ (streamable)
      description: >
        Clients sending `Accept: text/event-stream` receive a Server-Sent Events stream with one
        `event: section` per canonical section (§00 §8), followed by `event: done` carrying the
        persisted devotionId. Clients sending `Accept: application/json` receive the full Devotion
        object once generation completes. See §09 §6 for the full SSE protocol.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDevotionRequest'
      responses:
        "201":
          description: Devotion generated (non-streamed JSON response)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Devotion'
            text/event-stream:
              schema:
                type: string
                description: See §09 §6 — sequence of `section`/`done`/`error` SSE events
        "429":
          $ref: '#/components/responses/QuotaExceeded'
...
  /ai-conversations:
    get:
      tags: [AI Study]
      summary: List the authenticated user's AI Bible Study conversations
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AIConversation'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'
    post:
      tags: [AI Study]
      summary: Start a new conversation
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                contextPassageKey: { type: string, nullable: true, example: "ROM.8.28-39" }
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AIConversation'
  /ai-conversations/{id}:
    get:
      tags: [AI Study]
      summary: Get a conversation with its messages
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AIConversationDetail'
        "404":
          $ref: '#/components/responses/NotFound'
  /ai-conversations/{id}/messages:
    post:
      tags: [AI Study]
      summary: Ask a question in a conversation (streamable assistant reply)
      description: >
        Same SSE/JSON dual-mode contract as `/devotions/generate` (§09 §6). The assistant message
        is streamed token-by-token under `event: section` with `section: "answer"`, then
        `event: done` carries the persisted AIMessage id.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, example: "What does it mean that 'all things work together for good'?" }
                category:
                  $ref: '#/components/schemas/AIQuestionCategory'
      responses:
        "201":
          description: Assistant reply (non-streamed)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AIMessage'
            text/event-stream:
              schema:
                type: string
                description: See §09 §6
        "429":
          $ref: '#/components/responses/QuotaExceeded'
        "451":
          description: Blocked by content-safety policy (crisis-resource intercept, §12 §3)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ---------------------------------------------------------------------------------
  # AI PRAYER GENERATOR
  # ---------------------------------------------------------------------------------
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Search/list users (SUPPORT, ADMIN, SUPER_ADMIN)
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string, description: "Search by email, name, or ID" }
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: tier
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SubscriptionTier'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  meta:
                    $ref: '#/components/schemas/OffsetPaginationMeta'
        "403":
          $ref: '#/components/responses/Forbidden'
  /admin/users/{id}/role:
    patch:
      tags: [Admin]
      summary: Change a user's role (ADMIN, SUPER_ADMIN) — writes AdminAuditLog
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: '#/components/schemas/UserRole'
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        "403":
          $ref: '#/components/responses/Forbidden'
  /admin/organizations/{id}:
    patch:
      tags: [Admin]
      summary: Update an organization's seats/billing (ADMIN, SUPER_ADMIN)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                seatLimit: { type: integer }
                billingEmail: { type: string, format: email, nullable: true }
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Organization'
  /admin/ai-usage:
    get:
      tags: [Admin]
      summary: Aggregate AI usage analytics (ADMIN, SUPER_ADMIN)
      parameters:
        - name: feature
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AIFeature'
        - name: from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date }
      responses:
        "200":
          description: Aggregated usage + cost
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AIUsageSummary'
  /admin/moderation/images:
    get:
      tags: [Admin]
      summary: List flagged image generations (MODERATOR, ADMIN, SUPER_ADMIN)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        "200":
          description: Flagged images
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageGeneration'
                  meta:
                    $ref: '#/components/schemas/OffsetPaginationMeta'
  /admin/moderation/posts/{id}:
    patch:
      tags: [Admin]
      summary: Resolve a flagged community post (MODERATOR, ADMIN, SUPER_ADMIN)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [moderationStatus]
              properties:
                moderationStatus:
                  $ref: '#/components/schemas/ModerationStatus'
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CommunityPost'

# =====================================================================================
# COMPONENTS
# =====================================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Opaque cuid()
    CursorParam:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, maximum: 100 }
    PageParam:
      name: page
      in: query
      required: false
      schema: { type: integer, default: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, default: 20, maximum: 100 }
  responses:
...
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    QuotaExceeded:
      description: Tier-based AI usage quota exceeded (429 RATE_LIMITED, code=QUOTA_EXCEEDED in details)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string }
            details:
              type: object
              nullable: true
              additionalProperties: true
    CursorPaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            nextCursor: { type: string, nullable: true }
            hasMore: { type: boolean }
    OffsetPaginationMeta:
      type: object
      properties:
        pagination:
          type: object
          properties:
            page: { type: integer }
            pageSize: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    EnvelopeOrganization:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Organization'

    # --- Enums (mirror schema.prisma verbatim, §00 §13) ---------------------
    UserRole:
      type: string
      enum: [USER, GROUP_LEADER, MINISTRY_ADMIN, MODERATOR, CONTENT_EDITOR, SUPPORT, ADMIN, SUPER_ADMIN]
    SubscriptionTier:
      type: string
      enum: [FREE, PLUS, PREMIUM, MINISTRY, LIFETIME]
...
    DevotionType:
      type: string
      enum: [PERSONAL, FAMILY, YOUTH, CHILDREN, PATHFINDER, WOMEN, MEN, LEADERSHIP, MARRIAGE, EVANGELISTIC, SABBATH_SCHOOL, SMALL_GROUP, SERMON_PREP]
    DevotionDepth:
      type: string
      enum: [QUICK, STANDARD, DEEP, ADVANCED, TEACHING, SERMON]
...
    AIQuestionCategory:
      type: string
      enum: [MEANING, SIMPLE_EXPLANATION, CHILDRENS_EXPLANATION, HISTORICAL_BACKGROUND, CULTURAL_CONTEXT, MODERN_APPLICATION, HEBREW_ORIGINAL, GREEK_ORIGINAL, CROSS_REFERENCES, CHARACTER_PARALLELS, CHRIST_CONNECTION, GENERAL]
    MessageRole:
      type: string
      enum: [USER, ASSISTANT, SYSTEM]
    ImageStyle:
      type: string
      enum: [PHOTOREALISTIC, CINEMATIC, CHRISTIAN_ARTWORK, CARTOON, CHILDRENS_STORYBOOK, COMIC_BOOK, WATERCOLOR, OIL_PAINTING, MINIMALIST_POSTER, SOCIAL_MEDIA_GRAPHIC]
    ImageSourceType:
      type: string
      enum: [VERSE, CHAPTER, BIBLE_STORY, DEVOTION, PRAYER, SERMON, MEMORY_VERSE, CUSTOM_PROMPT]
    GenerationStatus:
      type: string
      enum: [PENDING, PROCESSING, COMPLETED, FAILED, MODERATION_BLOCKED]
...
    CommunityPostType:
      type: string
      enum: [DEVOTION_SHARE, TESTIMONY, PRAYER_REQUEST, INSIGHT, IMAGE_SHARE]
    ModerationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, FLAGGED]
...
    AIFeature:
      type: string
      enum: [DEVOTION, BIBLE_STUDY, PRAYER, SERMON, SCRIPTURE_ANALYSIS, IMAGE_GENERATION, STORY_VISUALIZER, CONTENT_CREATOR]

    # --- Core models ---------------------------------------------------------
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email, nullable: true }
        phone: { type: string, nullable: true }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
        role:
          $ref: '#/components/schemas/UserRole'
        tier:
          $ref: '#/components/schemas/SubscriptionTier'
        denominationLens: { type: string, nullable: true, example: "NEUTRAL" }
        locale: { type: string, example: "en" }
        timezone: { type: string, example: "UTC" }
        defaultBibleVersionId: { type: string, nullable: true }
        emailVerifiedAt: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Organization:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, example: "CHURCH" }
        seatLimit: { type: integer }
        billingEmail: { type: string, format: email, nullable: true }
        createdAt: { type: string, format: date-time }
...
    DevotionSections:
      type: object
      description: Canonical 15-section structured output, §00 §8
      properties:
        title: { type: string }
        keyScripture: { type: string }
        historicalContext: { type: string }
        biblicalContext: { type: string }
        verseExplanation: { type: string }
        theologicalInsights: { type: string }
        spiritualLessons: { type: string }
        lifeApplications: { type: string }
        reflectionQuestions:
          type: array
          items: { type: string }
        discussionQuestions:
          type: array
          items: { type: string }
        prayer: { type: string }
        actionSteps:
          type: array
          items: { type: string }
        memoryVerse: { type: string }
        relatedScriptures:
          type: array
          items: { type: string }
        closingEncouragement: { type: string }
    Devotion:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        passageKey: { type: string, example: "ROM.8.28-39" }
        versionCode: { type: string, example: "ESV" }
        type:
          $ref: '#/components/schemas/DevotionType'
        depth:
          $ref: '#/components/schemas/DevotionDepth'
        sections:
          $ref: '#/components/schemas/DevotionSections'
        memoryVerseRef: { type: string, nullable: true }
        isFavorite: { type: boolean }
        isShared: { type: boolean }
        createdAt: { type: string, format: date-time }
    GenerateDevotionRequest:
      type: object
      required: [passageKey, type, depth]
      properties:
        passageKey: { type: string, example: "ROM.8.28-39" }
        versionCode: { type: string, default: "ESV" }
        type:
          $ref: '#/components/schemas/DevotionType'
        depth:
          $ref: '#/components/schemas/DevotionDepth'
        themes:
          type: array
          items: { type: string }
          description: Theme slugs selected during Theme Discovery (§02 §1.3)

    # --- AI Bible Study Assistant™ ---------------------------------------------------------
    AIConversation:
      type: object
      properties:
        id: { type: string }
        title: { type: string, nullable: true }
        contextPassageKey: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AIConversationDetail:
      allOf:
        - $ref: '#/components/schemas/AIConversation'
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: '#/components/schemas/AIMessage'
    AIMessage:
      type: object
      properties:
        id: { type: string }
        conversationId: { type: string }
        role:
          $ref: '#/components/schemas/MessageRole'
        category:
          allOf:
            - $ref: '#/components/schemas/AIQuestionCategory'
          nullable: true
        content: { type: string }
        citations:
          type: array
          nullable: true
          items:
            type: object
            properties:
              ref: { type: string, example: "John 3:16" }
              versionCode: { type: string, example: "ESV" }
        createdAt: { type: string, format: date-time }

    # --- Prayer ---------------------------------------------------------
...
    ImageGeneration:
      type: object
      properties:
        id: { type: string }
        sourceType:
          $ref: '#/components/schemas/ImageSourceType'
        sourceKey: { type: string }
        style:
          $ref: '#/components/schemas/ImageStyle'
        refinedPrompt: { type: string }
        status:
          $ref: '#/components/schemas/GenerationStatus'
        resultUrl: { type: string, nullable: true }
        thumbnailUrl: { type: string, nullable: true }
        isWatermarked: { type: boolean }
        moderationStatus:
          $ref: '#/components/schemas/ModerationStatus'
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
...
    CommunityPost:
      type: object
      properties:
        id: { type: string }
        type:
          $ref: '#/components/schemas/CommunityPostType'
        content: { type: string }
        moderationStatus:
          $ref: '#/components/schemas/ModerationStatus'
        likeCount: { type: integer }
        commentCount: { type: integer }
        author:
          $ref: '#/components/schemas/User'
        createdAt: { type: string, format: date-time }
...
    AIUsageSummary:
      type: object
      properties:
        totalRequests: { type: integer }
        totalCostUsdMicros: { type: integer }
        byFeature:
          type: array
          items:
            type: object
            properties:
              feature:
                $ref: '#/components/schemas/AIFeature'
              requests: { type: integer }
              costUsdMicros: { type: integer }
              avgLatencyMs: { type: integer }
              errorRate: { type: number, format: float }